clean up clippy warnings and fix README inaccuracies
- README: fix threads comment (pipeline → multi-core), update rustls
  (0.23.37) and aws-lc-rs (1.16.2) versions
- proto: clone-on-copy, mem::replace → mem::take, KeyPairQueue type alias
- core: collapsible ifs, unnecessary u16 casts, range contains
- net: collapsible ifs, Box large enum variant, needless range loop,
  unnecessary i32 casts, int_plus_one

+++ quictun-core/src/icmp.rs
@@ -164,8 +164,8 @@ fn internet_checksum(data: &[u8]) -> u16 {
 
 /// RFC 1624 incremental checksum update.
 fn incremental_update(old_cksum: u16, old_val: u16, new_val: u16) -> u16 {
-    let mut sum: i32 = !(old_cksum) as u16 as i32;
-    sum += !(old_val) as u16 as i32;
+    let mut sum: i32 = (!old_cksum) as i32;
+    sum += (!old_val) as i32;
     sum += new_val as i32;
     while sum > 0xffff {
         sum = (sum & 0xffff) + (sum >> 16);
